Rename: 컴포넌트 파일 위치 변경
