+++ src/widgets/components/applications/applications.module.css
