+++ src/widgets/components/applications/index.tsx
