+++ src/widgets/components/dock/dock.module.css
