+++ src/widgets/components/dock/index.tsx
