+++ src/widgets/components/loader/index.tsx
