+++ src/widgets/components/loader/loader.module.css
