+++ src/widgets/components/loading/index.tsx
