+++ src/widgets/components/loading/loading.css
