+++ src/widgets/components/lock-screen/index.tsx
