+++ src/widgets/components/lock-screen/lock-screen.module.css
