+++ src/widgets/components/mobile-guide/index.tsx
