+++ src/widgets/components/mobile-guide/mobile-guide.module.css
