+++ src/widgets/components/navigation/index.tsx
@@ -1,6 +1,6 @@
 import styles from "./navigation.module.css";
 
-import { RealTime } from "@/entities/components";
+import { RealTime } from "@/widgets/components";
 import { useCheckOS } from "@/shared/hooks";
 
 const Navigation = () => {
@@ -10,12 +10,8 @@ const Navigation = () => {
     <div className={styles.container}>
       <div className={styles.leftArea}>
         <img className={styles.logo} src="../svgs/apple-logo.svg" />
-        <span className={`${styles.title} ${OS === "mac" ? styles.mac : ""}`}>
-          Finder
-        </span>
-        <span className={`${styles.item} ${OS === "mac" ? styles.mac : ""}`}>
-          File
-        </span>
+        <span className={`${styles.title} ${OS === "mac" ? styles.mac : ""}`}>Finder</span>
+        <span className={`${styles.item} ${OS === "mac" ? styles.mac : ""}`}>File</span>
       </div>
       <div className={styles.rightArea}>
         <RealTime />
