+++ src/widgets/components/outline/index.tsx
