+++ src/widgets/components/outline/outline.module.css
