+++ src/widgets/components/realtime/index.tsx
