+++ src/widgets/components/realtime/realtime.module.css
