+++ src/widgets/components/window/index.tsx
