+++ src/widgets/components/window/window.css
